+++ ProjectLighthouse.Servers.GameServer/Controllers/Resources/ResourcesController.cs
@@ -42,11 +42,11 @@ public class ResourcesController : ControllerBase
         string fullPath = Path.GetFullPath(path);
 
         // Prevent directory traversal attacks
-        if (!fullPath.StartsWith(FileHelper.FullResourcePath)) return this.StatusCode(400);
+        if (!fullPath.StartsWith(FileHelper.FullResourcePath)) return this.BadRequest();
 
         if (FileHelper.ResourceExists(hash)) return this.File(IOFile.OpenRead(path), "application/octet-stream");
 
-        return this.StatusCode(404);
+        return this.NotFound();
     }
 
     // TODO: check if this is a valid hash
